feature/frontend : adding a confirmation box before deleting an item

+++ FullStackApp/client-app/src/Department.js
@@ -80,23 +80,25 @@ export class Department extends Component {
       );
   }
   deleteClick(id) {
-    fetch(variables.API_URL + "department/" + id, {
-      method: "DELETE",
-      headers: {
-        Accept: "application/json",
-        "Content-Type": "application/json",
-      },
-    })
-      .then((res) => res.json())
-      .then(
-        (result) => {
-          alert(result);
-          this.refreshList();
+    if (window.confirm("Are you sure?")) {
+      fetch(variables.API_URL + "department/" + id, {
+        method: "DELETE",
+        headers: {
+          Accept: "application/json",
+          "Content-Type": "application/json",
         },
-        (error) => {
-          alert("failed");
-        }
-      );
+      })
+        .then((res) => res.json())
+        .then(
+          (result) => {
+            alert(result);
+            this.refreshList();
+          },
+          (error) => {
+            alert("failed");
+          }
+        );
+    }
   }
 
   refreshList() {
